new anisotropy works with constant viscosity model (cst only), try ShearTemplate.txt ;)
