+++ AnistropyUnitTests/data4plot.txt
@@ -1,0 +1,12 @@
+angle Txx Tzz Txz Tii
+0.000000 0.000000 0.000000 -0.250000 0.250000
+9.000000 -0.220419 0.220419 -0.321619 0.389902
+18.000000 -0.356646 0.356646 -0.509119 0.621609
+27.000000 -0.356646 0.356646 -0.740881 0.822254
+36.000000 -0.220419 0.220419 -0.928381 0.954189
+45.000000 -0.000000 0.000000 -1.000000 1.000000
+54.000000 0.220419 -0.220419 -0.928381 0.954189
+63.000000 0.356646 -0.356646 -0.740881 0.822254
+72.000000 0.356646 -0.356646 -0.509119 0.621609
+81.000000 0.220419 -0.220419 -0.321619 0.389902
+90.000000 0.000000 -0.000000 -0.250000 0.250000
