+++ AnistropyUnitTests/gnuplot_data.txt
@@ -1,0 +1,11 @@
+0.000000	-1.000000	0.000000 
+9.000000	-1.000000	0.000000 
+18.000000	-1.000000	0.000000 
+27.000000	-1.000000	0.000000 
+36.000000	-1.000000	0.000000 
+45.000000	-1.000000	0.000000 
+54.000000	-1.000000	0.000000 
+63.000000	-1.000000	0.000000 
+72.000000	-1.000000	0.000000 
+81.000000	-1.000000	0.000000 
+90.000000	-1.000000	0.000000
