+++ AnistropyUnitTests/visu.py
@@ -1,0 +1,3 @@
+import numpy as np
+with open('gnuplot_data.txt') as f:
+    lines = f.readlines()
